requirements.txt
# Core dependencies for the Ocean Plastic Sentinel
# requirements.txt

# Core dependencies for the Ocean Plastic Sentinel
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
numpy>=1.24.0

# Streamlit and visualization
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0

# Data handling and storage
geopy>=2.4.0

# Image processing for satellite data
pillow>=10.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0

# Data handling and storage  # Built-in to Python, but listed for clarity
requests>=2.31.0
numpy>=1.24.0
geopy>=2.4.0

# Image processing for satellite data
pillow>=10.0.0

# Async support
asyncio  # Built-in to Python

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: Enhanced functionality
# beautifulsoup4>=4.12.0  # For web scraping tools
# pandas>=2.0.0  # For data analysis tools
numpy>=1.24.0  # For numerical computations